+++ tests/services/analysis_starter/file_creators/test_gene_panel_file_creator.py
@@ -141,7 +141,7 @@ def test_gene_panel_file_creation_with_filtering(
     ]
 
     gene_panel_creator.create(
-        case_id=nallo_case_id, file_path=Path("nallo.bed"), header_filtering=True
+        case_id=nallo_case_id, file_path=Path("nallo.bed"), double_hashtag_filtering=True
     )
     write_mock.assert_called_once_with(
         content=expected_nallo_gene_panel_file_content, file_path=Path("nallo.bed")
